Add files via upload
Pololu 3pi+ 32U4 (see Fig. 1) was one of the classical programmable robots with two sensor systems, namely the ground light sensors and the front-end bump sensors. To realize the obstacle avoidance function, a collision sensing system could be critical which can be the only way to detect objects, since there was no image sensor or lidar in the robot.[2] The two bump sensors situated in the front of the 3pi+ robot which could measure off the reflected infra-red (IR) light from two plastic bumpers.  Each bump sensor could provide a feedback value. And once the plastic bumper at the front was pressed, the feed value would increase.[3] This allowed 3pi+ robot to detect the collision when it hit another object. The feedback from two bump sensors could determine the location of the object. Besides, the application of such sensors did not require a large amount of data calculation or image analysis, because the robot could detect the existence of obstacles in an intuitive way.  The 3pi+ robot had a rotary encode per motor which could provide a value according to the rotation of the wheel. With relationship between the practical distance and the encode value, it was possible that a coordinate system can be established using an estimate of robot position calculated by the accommodation of movement.[4] Therefore, the shape of the obstacles could be generated by activating the bump sensors continuously and the coordinate system.