Add files via upload
The aim of the project is to predict the number of rental bikes out at each hour based 
on the given previous data. The data presents the bike rent in 2017 with the weather 
condition such as temperature, humidity, wind speed and so on. Therefore, the 
project is a regression problem. Python based on Jupyter notebook was used to deal 
the problem. Baseline model can be created using the mean strategy based 
on 𝑑𝑢𝑚𝑚𝑦 𝑟𝑒𝑔𝑟𝑒𝑠𝑠𝑜𝑟. To solve a regression problem, linear regression model, 
decision tree, random forest can be built to predict the bike rent at each hour.